Partially working bouncy projectiles.
Edited the details for Mesh and CollisionComponent. Simulate Physics set to on, Simulation Generates Hit Events set to on, using new physics material. Collision preset has to be set to "BlockAll" for CollisionComponent for the projectile mesh to actually move. However, this means the mesh will collide with CollisionComponent, leading to it disappearing immediate if IsBouncy is not set. Also, Mesh and CollisionComponent (CC) will quickly separate. If Simulate Physics is only set for CC and not Mesh, then Mesh and CC stay together but the projectile will fly in the direction of the player. The muzzle was also moved forward to avoid collisions with the gun. Also, changing Collision Presets from BlockAll to Custom breaks the mesh's movement, even if all the settings remain the same.